requirements.txt
sqlalchemy>=1.4.0,<1.5.0
python-dotenv>=0.19.0,<0.20.0
vllm>=0.2.0
torch>=2.0.0
transformers>=4.30.0
Pillow>=9.0.0
opencv-python>=4.5.0
numpy>=1.21.0
python-multipart>=0.0.5
requests>=2.26.0
tqdm>=4.62.0
typing-extensions>=4.0.0
langchain-core
langgraph>0.2.27
langchain[openai]


fastapi~=0.111
uvicorn[standard]
pydantic~=2.7  
accelerate
paddleocr
ultralytics
whispercpp
TTS                       # Coqui‑TTS
python-multipart          # 若需上传文件